fix(vitepress): fix file names
