+++ docs/.vitepress/typescript/node/generateSidebar.ts
@@ -1,24 +1,77 @@
-import { HeadConfig } from 'vitepress'
+import glob from 'glob'
+import { DefaultTheme } from 'vitepress'
 
-const gtagHead: HeadConfig[] = <string>process.env.VITE_APP_GOOGLE_TAG_ID
-  ? ([
-      [
-        'script',
-        {
-          async: '',
-          src: `https://www.googletagmanager.com/gtag/js?id=${<string>(
-            process.env.VITE_APP_GOOGLE_TAG_ID
-          )}`,
-        },
-      ],
-      [
-        'script',
-        {},
-        `window.dataLayer = window.dataLayer || [];\nfunction gtag(){dataLayer.push(arguments);}\ngtag('js', new Date());\ngtag('config', '${<
-          string
-        >process.env.VITE_APP_GOOGLE_TAG_ID}');`,
-      ],
-    ] as HeadConfig[])
-  : []
+export default function generateSidebar() {
+  const sidebar: DefaultTheme.Sidebar = []
 
-export default gtagHead
+  glob
+    .sync('docs/**/*.md')
+    .sort((a, b) => {
+      const aArray = a.split('/')
+      const bArray = b.split('/')
+
+      for (let i = 0; i < Math.min(aArray.length, bArray.length); i++) {
+        const aName = aArray[i]
+        const bName = bArray[i]
+        if (aName !== bName) {
+          if (aName === 'index.md') {
+            return -1
+          }
+          if (bName === 'index.md') {
+            return 1
+          }
+          if (aName.endsWith('.md') && !bName.endsWith('.md')) {
+            return -1
+          }
+          if (!aName.endsWith('.md') && bName.endsWith('.md')) {
+            return 1
+          }
+          return aName.localeCompare(bName)
+        }
+      }
+
+      return 0
+    })
+    .map((path) => path.replace('docs/', '').replace(/_/g, ''))
+    .forEach((path) =>
+      path.split('/').forEach((text, index, array) => {
+        let items = sidebar
+
+        for (let i = 0; i < index; i++) {
+          items = (
+            items.find(
+              (child) => typeof child === 'object' && child.text === array[i]
+            ) as DefaultTheme.SidebarMulti
+          ).items
+        }
+
+        if (text === 'index.md') {
+          const pageName = array[index - 1]
+
+          items.push({
+            text: pageName ? `${pageName}介紹` : '首頁',
+            link: `/${path.replace('index.md', '')}`,
+          })
+          return
+        }
+
+        if (text.endsWith('.md')) {
+          items.push({
+            text: text.replace('.md', ''),
+            link: `/${path.replace('.md', '')}`,
+          })
+          return
+        }
+
+        const child = items.find(
+          (child) => typeof child === 'object' && child.text === text
+        ) as DefaultTheme.Sidebar
+
+        if (!child) {
+          items.push({ text: text, items: [], collapsed: true })
+        }
+      })
+    )
+
+  return sidebar
+}
